// autogpt_platform/frontend/src/app/api/__generated__/models/graphExecutionJobInfo.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { GraphExecutionJobInfoScheduleId } from "./graphExecutionJobInfoScheduleId";
import type { GraphExecutionJobInfoAgentName } from "./graphExecutionJobInfoAgentName";
import type { GraphExecutionJobInfoInputData } from "./graphExecutionJobInfoInputData";
import type { GraphExecutionJobInfoInputCredentials } from "./graphExecutionJobInfoInputCredentials";

export interface GraphExecutionJobInfo {
  schedule_id?: GraphExecutionJobInfoScheduleId;
  user_id: string;
  graph_id: string;
  graph_version: number;
  agent_name?: GraphExecutionJobInfoAgentName;
  cron: string;
  input_data: GraphExecutionJobInfoInputData;
  input_credentials?: GraphExecutionJobInfoInputCredentials;
  id: string;
  name: string;
  next_run_time: string;
  /** Timezone used for scheduling */
  timezone?: string;
}